+++ src/app.ts
@@ -1,0 +1,48 @@
+import express from "express";
+import { isProduction } from "./config/config.js";
+import { logger } from "./middleware/logger.js";
+import helmet from "helmet";
+import compression from "compression";
+import { engine } from "express-handlebars";
+import path from "path";
+import { fileURLToPath } from "url";
+import { Request, Response } from "express";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const app = express();
+
+if (isProduction) {
+  app.set("trust proxy", 1);
+}
+
+app.use(helmet());
+app.use(logger);
+app.use(compression());
+app.disable("x-powered-by");
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
+app.use(express.static("public"));
+
+app.engine(
+  "hbs",
+  engine({
+    extname: ".hbs",
+    defaultLayout: "main",
+    layoutsDir: path.join(__dirname, "../views/layouts"),
+    partialsDir: path.join(__dirname, "../views/partials"),
+    helpers: {
+      eq: (a: unknown, b: unknown) => a === b,
+    },
+  }),
+);
+
+app.set("view engine", "hbs");
+app.set("views", path.join(__dirname, "../views"));
+
+app.get("/", (req: Request, res: Response) => {
+  res.render("pages/board");
+});
+
+export default app;